requirements.txt
flask
openai
PyPDF2
sentence-transformers
faiss-cpu
sqlalchemy
python-dotenv
flask-limiter
flask-cors
tqdm
